requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
emoji==0.5.1
lxml==4.3.0
python-telegram-bot==11.1.0
requests==2.20.1
requests-toolbelt==0.9.1
